+++ openmls/src/group/mls_group/tests_and_kats/tests/create_commit_params.rs
@@ -1,5 +1,5 @@
 use crate::group::{
-    create_commit_params::CreateCommitParams,
+    create_commit::CreateCommitParams,
     mls_group::{FramingParameters, Proposal, WireFormat},
 };
