+++ openmls/src/group/tests_and_kats/tests/commit_validation.rs
@@ -1,7 +1,7 @@
 //! This module tests the validation of commits as defined in
 //! https://openmls.tech/book/message_validation.html#commit-message-validation
 
-use create_commit_params::CreateCommitParams;
+use create_commit::CreateCommitParams;
 use openmls_traits::{prelude::*, signatures::Signer, types::Ciphersuite};
 use proposal_store::QueuedProposal;
 use tls_codec::{Deserialize, Serialize};
